Merged 0.03 version to master

+++ lib/store/user/user_store.dart
@@ -1,17 +1,17 @@
 import 'package:mobx/mobx.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
-import 'package:watrix/models/local/last_activities.dart';
-import 'package:watrix/models/local/show_history.dart';
-import 'package:watrix/models/network/base_model.dart';
-import 'package:watrix/models/network/enums/entity_type.dart';
-import 'package:watrix/models/network/trakt/trakt_progress.dart';
-import 'package:watrix/models/network/cinenexa_user.dart';
-import 'package:watrix/models/network/trakt_user.dart';
-import 'package:watrix/models/network/user_stats.dart';
-import 'package:watrix/services/local/database.dart';
-import 'package:watrix/services/network/trakt_oauth_client.dart';
-import 'package:watrix/services/network/trakt_repository.dart';
-import 'package:watrix/store/favorites/favorites_store.dart';
+import 'package:cinenexa/models/local/last_activities.dart';
+import 'package:cinenexa/models/local/show_history.dart';
+import 'package:cinenexa/models/network/base_model.dart';
+import 'package:cinenexa/models/network/enums/entity_type.dart';
+import 'package:cinenexa/models/network/trakt/trakt_progress.dart';
+import 'package:cinenexa/models/network/cinenexa_user.dart';
+import 'package:cinenexa/models/network/trakt_user.dart';
+import 'package:cinenexa/models/network/user_stats.dart';
+import 'package:cinenexa/services/local/database.dart';
+import 'package:cinenexa/services/network/trakt_oauth_client.dart';
+import 'package:cinenexa/services/network/trakt_repository.dart';
+import 'package:cinenexa/store/favorites/favorites_store.dart';
 part 'user_store.g.dart';
 
 class UserStore = _UserStoreBase with _$UserStore;
@@ -75,17 +75,17 @@ abstract class _UserStoreBase with Store {
 
       List<Future> listFutures = [];
 
-      if (localLast != null) {
-        if (lastActivities.epWatchedAt.isAfter(localLast.epWatchedAt)) {
+      if (localLast != null && localLast.movieCollectedAt != null) {
+        if (lastActivities.epWatchedAt!.isAfter(localLast.epWatchedAt!)) {
           listFutures.add(fetchUserWatchedShows(
               api: lastActivities, local: localLast, fromApi: true));
         } else {
           listFutures.add(
               fetchUserWatchedShows(api: lastActivities, local: localLast));
         }
-        if (lastActivities.movieCollectedAt
-                .isAfter(localLast.movieCollectedAt) ||
-            lastActivities.epCollectedAt.isAfter(localLast.epCollectedAt)) {
+        if (lastActivities.movieCollectedAt!
+                .isAfter(localLast.movieCollectedAt!) ||
+            lastActivities.epCollectedAt!.isAfter(localLast.epCollectedAt!)) {
           listFutures.add(
               favoritesStore?.fetchFavorites(fromApi: true) ?? Future.value());
         } else {
@@ -100,8 +100,7 @@ abstract class _UserStoreBase with Store {
       }
       Future.wait(listFutures).whenComplete(() => localDb.addLastActivities(
           lastActivities: lastActivities
-            ..extensionsSyncedAt =
-                localLast?.extensionsSyncedAt ?? DateTime.now()));
+            ..extensionsSyncedAt = localLast?.extensionsSyncedAt));
     }
   }
 
@@ -136,6 +135,10 @@ abstract class _UserStoreBase with Store {
   }
 
   Future _removeProgressFromApi(TraktProgress traktProgress) async {
+    if (!isTraktLogged) {
+      return;
+    }
+
     if (traktProgress.playbackId != null) {
       return repository.removeProgress(progressId: traktProgress.playbackId!);
     }
